requirements.txt
django==5.0.3
django-rest-framework==0.1.0
django-countries==7.5.1
django-rest-swagger==2.2.0

